+++ .gitignore
@@ -1,5 +1,6 @@
 # Build artifacts.
 /bin/
+/cmd/oc-mirror/data/oc-mirror-v2
 
 # Test artifacts.
 coverage.out
