+++ Makefile
@@ -18,10 +18,8 @@ include $(addprefix ${GOMODCACHE}/${BUILD_MACHINERY_PATH}@${BUILD_MACHINERY_VERS
 	targets/openshift/deps-gomod.mk \
 )
 
-GO_LD_EXTRAFLAGS=$(call version-ldflags,github.com/openshift/oc-mirror/v2/internal/pkg/version)
-
 GO_MOD_FLAGS = -mod=readonly
-GO_BUILD_PACKAGES := ./cmd/...
+GO_BUILD_PACKAGES := ./cmd/oc-mirror
 GO_PACKAGE = github.com/openshift/oc-mirror
 
 LIBDM_BUILD_TAG = $(shell hack/libdm_tag.sh)
@@ -62,12 +60,15 @@ hack-build: clean
 .PHONY: hack-build
 
 tidy:
-	$(GO) work sync
+	$(GO) mod tidy
+	cd v2 && $(GO) mod tidy
 .PHONY: tidy
 
 clean:
+	@rm -f cmd/oc-mirror/oc-mirror-v2
 	@rm -rf ./$(GO_BUILD_BINDIR)/*
 	@cd test/integration && make clean
+	make -C v2 clean
 .PHONY: clean
 
 test-unit:
@@ -115,7 +116,9 @@ vet:
 	cd v2 && $(GO) vet $(GO_MOD_FLAGS) $(GO_BUILD_FLAGS) ./...
 .PHONY: vet
 
-build: 
+build:
+	make -C v2 build
+	@cp v2/build/oc-mirror ./cmd/oc-mirror/data/oc-mirror-v2
 	mkdir -p $(GO_BUILD_BINDIR)
 	go build $(GO_MOD_FLAGS) $(GO_BUILD_FLAGS) $(GO_LD_FLAGS) -o $(GO_BUILD_BINDIR) ./...
 .PHONY: build
