+++ cmd/oc-mirror/main.go
@@ -1,21 +1,76 @@
 package main
 
 import (
+	"bytes"
+	"embed"
+	"errors"
+	"fmt"
+	"io"
 	"os"
+	"os/exec"
+	"path/filepath"
 	"slices"
+	"syscall"
 
+	"k8s.io/klog"
 	kcmdutil "k8s.io/kubectl/pkg/cmd/util"
 
 	cliV1 "github.com/openshift/oc-mirror/pkg/cli/mirror"
-	cliV2 "github.com/openshift/oc-mirror/v2/pkg/cli"
 )
+
+//go:embed data/*
+var mirrorV2 embed.FS
 
 func main() {
 	if slices.Contains(os.Args, "--v2") {
-		err := cliV2.RunOcMirrorV2()
-		kcmdutil.CheckErr(err)
+		err := runOcMirrorV2(os.Args)
+		var exitErr *exec.ExitError
+		if err != nil && errors.As(err, &exitErr) {
+			os.Exit(exitErr.ExitCode())
+		}
 	} else {
 		rootCmd := cliV1.NewMirrorCmd()
 		kcmdutil.CheckErr(rootCmd.Execute())
 	}
+}
+
+func runOcMirrorV2(args []string) error {
+	tmpdir, err := os.MkdirTemp("", "oc-mirror-")
+	if err != nil {
+		return fmt.Errorf("failed to create temp dir: %w", err)
+	}
+	defer os.RemoveAll(tmpdir)
+
+	path := filepath.Join(tmpdir, "oc-mirror")
+	klog.V(5).Infof("Unpacking v2 binary to %s", path)
+
+	binaryV2, err := mirrorV2.ReadFile("data/oc-mirror-v2")
+	if err != nil {
+		return fmt.Errorf("failed to read v2 binary: %w", err)
+	}
+
+	v2File, err := os.OpenFile(path, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0o755)
+	if err != nil {
+		return fmt.Errorf("failed to create v2 binary: %w", err)
+	}
+	if _, err := io.Copy(v2File, bytes.NewReader(binaryV2)); err != nil {
+		v2File.Close()
+		return fmt.Errorf("failed to write v2 binary: %w", err)
+	}
+	// We must close the binary file before we can execute it
+	v2File.Close()
+
+	cmd := exec.Cmd{
+		Path: path,
+		SysProcAttr: &syscall.SysProcAttr{
+			// Kill children if parent is dead
+			Pdeathsig: syscall.SIGKILL,
+			Setpgid:   true,
+		},
+		Args:   args,
+		Stdout: os.Stdout,
+		Stderr: os.Stderr,
+	}
+
+	return cmd.Run()
 }
