--- go.work
@@ -1,6 +1,0 @@
-go 1.23.0
-
-use (
-	.
-	./v2
-)
