+++ v2/Makefile
@@ -1,14 +1,50 @@
+GO := go
+
+ifdef OS_GIT_VERSION
+	SOURCE_GIT_TAG := ${OS_GIT_VERSION}
+endif
+
+GOMODCACHE=$(shell $(GO) env GOMODCACHE)
+
+BUILD_MACHINERY_VERSION=v0.0.0-20240419090851-af9c868bcf52
+BUILD_MACHINERY_PATH=github.com/openshift/build-machinery-go
+
+$(shell $(GO) get ${BUILD_MACHINERY_PATH}@${BUILD_MACHINERY_VERSION})
+
+# Include the library makefile
+include $(addprefix ${GOMODCACHE}/${BUILD_MACHINERY_PATH}@${BUILD_MACHINERY_VERSION}/make/, \
+	golang.mk \
+	targets/openshift/images.mk \
+	targets/openshift/deps-gomod.mk \
+)
+
+GO_LD_EXTRAFLAGS=$(call version-ldflags,github.com/openshift/oc-mirror/v2/internal/pkg/version)
+
 .PHONY: all test build clean
 
 REGISTRY_BASE ?= quay.io/luzuccar
 IMAGE_NAME ?= golang-oci-mirror
 IMAGE_VERSION ?= v0.0.1
 
+GO_BUILD_BINDIR := ./build
+GO_MOD_FLAGS = -mod=readonly
+GO_BUILD_PACKAGES := ./cmd/oc-mirror
+GO_PACKAGE = github.com/openshift/oc-mirror/v2
+
+LIBDM_BUILD_TAG = $(shell ../hack/libdm_tag.sh)
+LIBSUBID_BUILD_TAG = $(shell ../hack/libsubid_tag.sh)
+BTRFS_BUILD_TAG = $(shell ../hack/btrfs_tag.sh) $(shell ../hack/btrfs_installed_tag.sh)
+
+ifeq ($(DISABLE_CGO), 1)
+	override BTRFS_BUILD_TAG = exclude_graphdriver_devicemapper exclude_graphdriver_btrfs containers_image_openpgp
+endif
+
+GO_BUILD_FLAGS = -tags "json1 $(BTRFS_BUILD_TAG) $(LIBDM_BUILD_TAG) $(LIBSUBID_BUILD_TAG)"
 all: clean test build
 
 build: 
-	mkdir -p build
-	go build -o build ./... 
+	mkdir -p $(GO_BUILD_BINDIR)
+	go build $(GO_MOD_FLAGS) $(GO_BUILD_FLAGS) $(GO_LD_FLAGS) -o $(GO_BUILD_BINDIR) ./... 
 	
 
 build-dev:
